new_requirement_20201223.txt
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
cryptography==3.3.1
cycler==0.10.0
Django==2.2.13
django-ckeditor==6.0.0
django-haystack==3.0
django-js-asset==1.2.2
djangorestframework==3.9.2
django-cors-headers==3.2.1
django-rest-swagger==2.1.2
geoip2==4.1.0
idna==2.10
idna-ssl==1.1.0
jieba==0.42.1
kiwisolver==1.3.1
matplotlib==3.3.3
maxminddb==2.0.3
multidict==5.1.0
numpy==1.19.4
Pillow==8.0.1
pycparser==2.20
PyMySQL==0.10.1
pyparsing==2.4.7
PyQt5==5.15.2
PyQt5-sip==12.8.1
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
requests==2.25.1
selenium==3.141.0
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
urllib3==1.26.2
uWSGI==2.0.19.1
Whoosh==2.7.4
wordcloud==1.8.1
yarl==1.6.3
